+++ workflows/index.ts
@@ -21,6 +21,7 @@ export function concreteWorkflowToMeta(
 
 export * from "./src/xstate.ts";
 export * from "./src/copy-template-machine/copy-template-machine.ts";
+export * from "./src/copy-template-machine/types.ts";
 export * from "./src/test-runner.ts";
 export * from "./src/chain.ts";
 export * from "./src/npm.ts";
